Remove unnecessary boxing and unboxing

+++ build-monitor-acceptance/src/main/java/com/smartcodeltd/aether/ConsoleTransferListener.java
@@ -55,13 +55,13 @@ public class ConsoleTransferListener
     @Override
     public void transferProgressed(TransferEvent event) {
         TransferResource resource = event.getResource();
-        downloads.put(resource, Long.valueOf(event.getTransferredBytes()));
+        downloads.put(resource, event.getTransferredBytes());
 
         StringBuilder buffer = new StringBuilder(64);
 
         for (Map.Entry<TransferResource, Long> entry : downloads.entrySet()) {
             long total = entry.getKey().getContentLength();
-            long complete = entry.getValue().longValue();
+            long complete = entry.getValue();
 
             buffer.append(getStatus(complete, total)).append("  ");
         }
